% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAltModeTrips.R
\docType{data}
\name{AltModeModels_ls}
\alias{AltModeModels_ls}
\title{Alternative model trip models}
\format{A list having 'Metro' and 'NonMetro' components. Each component has
'Walk', 'Bike' and 'Transit' components.}
\source{
CalcAltModeTrips.R script.
}
\usage{
AltModeModels_ls
}
\description{
A list of string representations for all alternative mode trip models
}
\keyword{datasets}

