% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAltModeTrips.R
\name{CalcAltModeTrips}
\alias{CalcAltModeTrips}
\title{Main module function to calculate alternative mode trips}
\usage{
CalcAltModeTrips(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalcAltModeTrips} calculates alternative mode trips for each
household.
}
\details{
This function calculates walk, bike, and transit tripss for households.
}

