% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAltModeTrips.R
\name{estimateHurdleTripModel}
\alias{estimateHurdleTripModel}
\title{Estimate hurdle trip model and return string representation}
\usage{
estimateHurdleTripModel(Formula, Data_df)
}
\arguments{
\item{Formula}{A model formula.}

\item{Data_df}{A data frame containing the estimation data}
}
\value{
A list containing two named string components. The 'Count' component
contains a string representation of the count model. The 'Zero' component
contains a string representation of the zero model.
}
\description{
\code{estimateHurdleTripModel} estimates a hurdle trip model and returns a
string representation of that model
}
\details{
This function estimates a hurdle model of trips for a specified mode
(transit, walk, bike) and development type. It returns a string
representation of the model.
}

